requirements.txt
streamlit>=1.30.0
playwright>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.27.0
openai>=1.30.0
pymupdf>=1.23.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
slowapi>=0.1.9
